assignment5

+++ public/js/inventory.js
@@ -33,13 +33,26 @@ async function buildInventoryList(data) {
         dataTable += '<tbody>'; 
         // Iterate over all vehicles in the array and put each in a row 
         data.forEach(function (element) { 
-            console.log(element.inv_id + ", " + element.inv_model); 
-            dataTable += `<tr><td>${element.inv_make} ${element.inv_model}</td>`; 
-            dataTable += `<td><a href='/inv/edit/${element.inv_id}' title='Click to update'>Modify</a></td>`; 
-            dataTable += `<td><a href='/inv/delete/${element.inv_id}' title='Click to delete'>Delete</a></td></tr>`; 
+            if (data) {
+                console.log(element.inv_id + ", " + element.inv_model); 
+                dataTable += `<tr><td>${element.inv_make} ${element.inv_model}</td>`; 
+                dataTable += `<td><a href='/inv/edit/${element.inv_id}' title='Click to update'>Modify</a></td>`; 
+                dataTable += `<td><a href='/inv/delete/${element.inv_id}' title='Click to delete'>Delete</a></td></tr>`; 
+            } else {
+                dataTable += `<tr><td>Nothing here for Now</td>`; 
+                dataTable += `<td><a href='#' title='Click to update'>Modify</a></td>`; 
+                dataTable += `<td><a href='#' title='Click to delete'>Delete</a></td></tr>`;
+            }
         }) 
         dataTable += '</tbody>'; 
         // Display the contents in the Inventory Management view 
         inventoryDisplay.innerHTML = dataTable; 
     }
 }
+
+
+const form = document.querySelector("#editinventoryform")
+form.addEventListener("change", function () {
+    const updateBtn = document.querySelector("button")
+    updateBtn.removeAttribute("disabled")
+})
